move s3 client util to its own file
