+++ ohmg/core/utils/__init__.py
@@ -36,22 +36,6 @@ def random_alnum(size=6):
     return code
 
 
-def get_boto3_s3_client():
-    import boto3
-
-    if settings.AWS_ACCESS_KEY_ID:
-
-        return boto3.client(
-            "s3",
-            region_name=settings.AWS_S3_REGION_NAME,
-            aws_access_key_id=settings.AWS_ACCESS_KEY_ID,
-            aws_secret_access_key=settings.AWS_SECRET_ACCESS_KEY,
-            endpoint_url=settings.AWS_S3_ENDPOINT_URL,
-        )
-    else:
-        return None
-
-
 MONTH_CHOICES = [
     (1, "JAN."),
     (2, "FEB."),
